Add CATALINA_OPTS to Dockerfiles
Catalina is the HTTP server component of Tomcat. DSpace docs say we
should make sure all Java components are running in UTF-8 mode, and
setting the heap to 512M is just to be verbose (as this seems to be
default, but better to have it explicitly so we can find it later).