Wait for PostgreSQL to become available
The DSpace container will attempt to create a PostgreSQL user and
database schema if needed, but we need to make sure the database is
actually available first. There are a few ways to do this, but I'll
just use the `wait-for-postgres.sh` example given in Docker's docs.

This eliminates race conditions when starting DSpace via a Docker
Compose environment.

See: https://docs.docker.com/compose/startup-order/