Move banner and bashrc back to root user
It was nice to have the default user for `docker exec` to be dspace
but we lost that perk when we finally figured out how to get the
cron service running—the side effect being that now `docker exec`
uses the root user. As such, it only makes sense to move these
configs back to root.