rootfs/dspace/bin/start-dspace: Rework privileges
For some reason we can't `su - dspace` here. The effective user is
always root. To get around this we can use `su -c` to run only the
commands we need as the dspace user. Note that this means we lose
the PATH that contains the catalina.sh script so we need to call
it with the full path.