Add support for hostname and port build args
We need to be able to adjust the build depending on the target env,
for example testing on a local developer's machine versus running
in production behind an nginx reverse proxy. In the first case it's
ok to run on port 8080, but in the second you'd rather run on port
80 or 443 and then redirect internally to port 8080, in which case
Tomcat would need to know via which port it is being proxied.