Dockerfile: Don't do a system upgrade
The Dockerfile best practices guide says to not do a system upgrade
during the build step. Removing this simplifies the Dockerfile and
I suppose that any gains from running this here would be due to the
upstream Dockerfile being out of date, which is actually not a real
concern right now (a better concern would be that we don't redeploy
our containers in production very often).

See:
https://docs.docker.com/engine/userguide/eng-image/dockerfile_best-practices/