Move helper scripts to /usr/local/sbin
These helper scripts don't belong to DSpace and the nature of Docker's
COPY instruction makes it so that target directories get overwritten,
meaning that the order of installing DSpace with ant and copying this
rootfs can become tricky if you're not paying attention.

This path is already in the shell's PATH so it's harmless to move the
scripts there. Also, this enables me to do multi-stage builds cleaner
because I can copy the build environment to /dspace and then COPY the
rootfs to / without them conflicting.